refactored UI
